Refactored exception package
