+++ src/main/java/tudelft/ti2806/pl3/Application.java
@@ -4,7 +4,7 @@ import newick.NewickParser;
 import tudelft.ti2806.pl3.controls.KeyController;
 import tudelft.ti2806.pl3.controls.WindowController;
 import tudelft.ti2806.pl3.data.graph.GraphDataRepository;
-import tudelft.ti2806.pl3.exeption.FileSelectorException;
+import tudelft.ti2806.pl3.exception.FileSelectorException;
 import tudelft.ti2806.pl3.sidebar.SideBarController;
 import tudelft.ti2806.pl3.util.FileSelector;
 import tudelft.ti2806.pl3.util.TreeParser;
